// src/rest/card/models/CardResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CardResponse
 */
export interface CardResponse {
    /**
     * 
     * @type {string}
     * @memberof CardResponse
     */
    number: string;
    /**
     * 
     * @type {string}
     * @memberof CardResponse
     */
    iban: string;
    /**
     * 
     * @type {number}
     * @memberof CardResponse
     */
    transactionCeiling: number;
    /**
     * 
     * @type {number}
     * @memberof CardResponse
     */
    rolling30Ceiling: number;
    /**
     * 
     * @type {boolean}
     * @memberof CardResponse
     */
    isActive?: boolean;
}

/**
 * Check if a given object implements the CardResponse interface.
 */
export function instanceOfCardResponse(value: object): value is CardResponse {
    if (!('number' in value) || value['number'] === undefined) return false;
    if (!('iban' in value) || value['iban'] === undefined) return false;
    if (!('transactionCeiling' in value) || value['transactionCeiling'] === undefined) return false;
    if (!('rolling30Ceiling' in value) || value['rolling30Ceiling'] === undefined) return false;
    return true;
}

export function CardResponseFromJSON(json: any): CardResponse {
    return CardResponseFromJSONTyped(json, false);
}

export function CardResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CardResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'number': json['number'],
        'iban': json['iban'],
        'transactionCeiling': json['transactionCeiling'],
        'rolling30Ceiling': json['rolling30Ceiling'],
        'isActive': json['isActive'] == null ? undefined : json['isActive'],
    };
}

export function CardResponseToJSON(json: any): CardResponse {
    return CardResponseToJSONTyped(json, false);
}

export function CardResponseToJSONTyped(value?: CardResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'number': value['number'],
        'iban': value['iban'],
        'transactionCeiling': value['transactionCeiling'],
        'rolling30Ceiling': value['rolling30Ceiling'],
        'isActive': value['isActive'],
    };
}